cache/gcs: rework CustomTime check/update logic
We used to get the attributes by touch (bumping CustomTime), but it seems that
that stresses GCS too much. Instead since GCS only supports day-level policies,
round to the next day, and bump the CustomTime only if the existing one if
different than what it should be. Concurrent updates are fine and should now
represent the vast minority of updates. However, it means that setting a TTL to
1 day could result in weird behaviour.

If for some reason we can't update the time, since of object exists consider
it's fine  because, in either way, worst case:
1. we return the error and bazel will re-executes the action
2. we don't return the error and the object will maybe be garbage
   collected before it should, in which case it'll come back to 1.

Let's be optimistic and chose 2 then.

Signed-off-by: Steeve Morin <steeve@zen.ly>